Adding codegen cmd for unilogin soap ws
