+++ package.json
@@ -12,7 +12,8 @@
     "codegen:covers": "rm -rf src/lib/rest/cover-service-api/generated/model/*.* && orval --project coverService",
     "codegen:graphql": "NODE_TLS_REJECT_UNAUTHORIZED=0 graphql-codegen --require tsconfig-paths/register --config codegen.ts",
     "codegen:publizon": "rm -rf lib/rest/publizon-api/generated/model/*.* && orval --project publizonAdapter",
-    "codegen:pubhub": "rm -rf lib/soap/publizon/v2_7/generated/* &&  npx node --loader ts-node/esm --no-warnings=ExperimentalWarning ./scripts/wsdl-tsclient/wsdl-codegen.ts",
+    "codegen:pubhub": "rm -rf lib/soap/publizon/v2_7/generated/* && npx tsx ./scripts/wsdl-tsclient/publizon-wsdl-codegen",
+    "codegen:unilogin": "rm -rf lib/soap/unilogin/wsiinst-v5/generated/* && npx tsx ./scripts/wsdl-tsclient/unilogin-wsdl-codegen",
     "dev:https": "NEXT_PUBLIC_APP_URL=https://dapple-cms.docker:3000 NODE_TLS_REJECT_UNAUTHORIZED=0 next dev -H dapple-cms.docker --experimental-https",
     "dev": "NODE_TLS_REJECT_UNAUTHORIZED=0 next dev",
     "format:check": "prettier --check .",
@@ -125,6 +126,7 @@
     "prettier-plugin-tailwindcss": "^0.6.11",
     "storybook": "^8.5.3",
     "tailwindcss": "^4.0.7",
+    "tsx": "^4.19.3",
     "typescript": "^5.6.2",
     "vite-tsconfig-paths": "^5.0.1",
     "vitest": "^2.1.3"
