+++ scripts/wsdl-tsclient/unilogin-wsdl-codegen.ts
@@ -1,0 +1,4 @@
+import { generateWsdlFiles } from "./helper"
+
+const wsdlFiles = ["./lib/soap/unilogin/wsiinst-v5/wsdl/ws.wsdl"]
+generateWsdlFiles(wsdlFiles)
